Finite water almost working

+++ Assets/Scripts/FiniteWaterDispersionRelation.hlsl
@@ -3,29 +3,29 @@
 
 #include "HLSLMath.hlsl"
 
-double FiniteWaterDispersionRelationDummy(double k, double fnh, double h, double theta)
+float FiniteWaterDispersionRelationDummy(float k, float fnh, float h, float theta)
 {
     // The 'k' provided is a guess, used to compute a new and updated value for 'k'.
-    return (double) fnh * (double) fnh * (double) k * (double) h * (double) cos((float) theta) * (double) cos((float) theta) - (double) CustomTanh((float) k * h);
+    return (float) fnh * (float) fnh * (float) k * (float) h * (float) cos((float) theta) * (float) cos((float) theta) - (float) CustomTanh((float) k * h);
 }
 
-double FiniteWaterDispersionRelationDerivativeDummy(double k, double fnh, double h, double theta)
+float FiniteWaterDispersionRelationDerivativeDummy(float k, float fnh, float h, float theta)
 {
     // k is a guess, used to compute a new k. Central differences are used to approximate the derivative.
-    return ((double) FiniteWaterDispersionRelationDummy((double) k + (double) 1e-3, (double) fnh, (double) h, (double) theta) - (double) FiniteWaterDispersionRelationDummy((double) k - (double) 1e-3, (double) fnh, (double) h, (double) theta)) / (double) 2e-3;
+    return ((float) FiniteWaterDispersionRelationDummy((float) k + (float) 1e-3, (float) fnh, (float) h, (float) theta) - (float) FiniteWaterDispersionRelationDummy((float) k - (float) 1e-3, (float) fnh, (float) h, (float) theta)) / (float) 2e-3;
 }
 
-double FiniteWaterDispersionRelation(double fnh, double h, double theta)
+float FiniteWaterDispersionRelation(float fnh, float h, float theta)
 {
     // The dispersion relation is solved using Newton's method. Settings are:
     int max_iter = 50;
-    double epsilon = 1e-10;
+    float epsilon = 1e-10;
 
     // Initialize for use in Newton's method.
-    double fxn, dfxn;
+    float fxn, dfxn;
 
     // Define initial guess for k, herby denoted as x.
-    double xn = 10.0;
+    float xn = 10.0;
 
     for (int n = 0; n < max_iter; n++)
     {
